Implement Serializers::StringSerializer.
- Extract Serializers::AttributesSerializer.
