+++ lib/serializers.rb
@@ -95,5 +95,6 @@ module Serializers
   autoload :BookSerializer,     'serializers/book_serializer'
   autoload :MechanicSerializer, 'serializers/mechanic_serializer'
   autoload :SourceSerializer,   'serializers/source_serializer'
+  autoload :StringSerializer,   'serializers/string_serializer'
   autoload :SpellSerializer,    'serializers/spell_serializer'
 end
