+++ lib/serializers/base_serializer.rb
@@ -7,61 +7,12 @@ require 'serializers'
 module Serializers
   # Abstract base class for serializing an object to a JSON-compatible hash.
   class BaseSerializer
-    class << self
-      def attribute(attr_name)
-        guard_abstract_class!
-
-        self::Attributes.send :def_delegator, :object, attr_name
-
-        attribute_definitions[attr_name.to_s] = -> { send(attr_name) }
-      end
-
-      def attributes(*attr_names)
-        guard_abstract_class!
-
-        attr_names.each { |attr_name| attribute(attr_name) }
-      end
-
-      protected
-
-      def defined_attributes
-        return attribute_definitions.dup unless superclass < BaseSerializer
-
-        superclass.defined_attributes.merge(attribute_definitions)
-      end
-
-      private
-
-      def attribute_definitions
-        @attribute_definitions ||= {}
-      end
-
-      def guard_abstract_class!
-        return unless self == Serializers::BaseSerializer
-
-        message =
-          'BaseSerializer is an abstract class and cannot define attributes.'
-
-        raise RuntimeError, message, caller[1..-1]
-      end
-
-      def inherited(subclass)
-        super
-
-        subclass.const_set(:Attributes, Module.new { extend Forwardable })
-        subclass.include(subclass::Attributes)
-      end
-    end
-
     def call(object)
       ensure_can_serialize!(object)
 
-      @object    = object
-      attributes = self.class.send(:defined_attributes)
+      @object = object
 
-      attributes.map.with_object({}) do |(attr_name, block), hsh|
-        hsh[attr_name] = instance_exec(&block)
-      end
+      serialize_object(object)
     end
     alias_method :serialize, :call
 
@@ -81,6 +32,11 @@ module Serializers
           object:     object,
           serializer: self
         )
+    end
+
+    def serialize_object(object)
+      raise Serializers::UndefinedSerializerError,
+        Serializers::UndefinedSerializerError.message(object)
     end
   end
 end
