+++ spec/serializers/base_serializer_spec.rb
@@ -4,100 +4,27 @@ require 'rails_helper'
 
 require 'serializers/base_serializer'
 
-require 'support/examples/serializer_examples'
-
 RSpec.describe Serializers::BaseSerializer do
-  include Spec::Support::Examples::SerializerExamples
-
-  shared_context 'with a serializer subclass' do
-    let(:described_class) { Spec::WidgetSerializer }
-    let(:object)          { Spec::Widget.new('Gadget', 'Medium', 'High') }
-
-    example_class 'Spec::Widget', Struct.new(:name, :size, :complexity)
-
-    # rubocop:disable RSpec/DescribedClass
-    example_class 'Spec::WidgetSerializer', Serializers::BaseSerializer
-    # rubocop:enable RSpec/DescribedClass
-  end
-
   subject(:serializer) { described_class.new }
 
-  describe '::new' do
+  describe '.new' do
     it { expect(described_class).to be_constructible.with(0).arguments }
   end
 
-  describe '::attribute' do
-    let(:error_message) do
-      'BaseSerializer is an abstract class and cannot define attributes.'
-    end
-
-    it { expect(described_class).to respond_to(:attribute).with(1).argument }
-
-    it 'should raise an error' do
-      expect { described_class.attribute :name }
-        .to raise_error RuntimeError, error_message
-    end
-
-    wrap_context 'with a serializer subclass' do
-      it 'should not raise an error' do
-        expect { described_class.attribute :name }.not_to raise_error
-      end
-
-      describe 'with an attribute name' do
-        before(:example) { described_class.attribute :name }
-
-        include_examples 'should serialize attributes', :name
-      end
-    end
-  end
-
-  describe '::attributes' do
-    let(:error_message) do
-      'BaseSerializer is an abstract class and cannot define attributes.'
-    end
-
-    it 'should define the method' do
-      expect(described_class)
-        .to respond_to(:attributes)
-        .with_unlimited_arguments
-    end
-
-    it 'should raise an error' do
-      expect { described_class.attributes :name, :size, :complexity }
-        .to raise_error RuntimeError, error_message
-    end
-
-    wrap_context 'with a serializer subclass' do
-      it 'should not raise an error' do
-        expect { described_class.attributes :name, :size, :complexity }
-          .not_to raise_error
-      end
-
-      describe 'with one attribute name' do
-        before(:example) { described_class.attributes :name }
-
-        include_examples 'should serialize attributes', :name
-      end
-
-      describe 'with many attribute names' do
-        before(:example) do
-          described_class.attributes :name, :size, :complexity
-        end
-
-        include_examples 'should serialize attributes',
-          :name,
-          :size,
-          :complexity
-      end
-    end
-  end
-
   describe '#call' do
+    let(:object) { Object.new.freeze }
+    let(:error_message) do
+      Serializers::UndefinedSerializerError.message(object)
+    end
+
     it { expect(serializer).to respond_to(:call).with(1).argument }
 
     it { expect(serializer).to alias_method(:call).as(:serialize) }
 
-    it { expect(serializer.call Object.new).to be == {} }
+    it 'should raise an error' do
+      expect { serializer.call object }
+        .to raise_error Serializers::UndefinedSerializerError, error_message
+    end
 
     context 'when the class defines can_serialize?' do
       let(:described_class) { Spec::WidgetSerializer }
@@ -119,98 +46,34 @@ RSpec.describe Serializers::BaseSerializer do
           'Unable to serialize nil with Spec::WidgetSerializer'
         end
 
-        it 'should raise an error' do
+        it 'should raise an invalid object error' do
           expect { serializer.call(nil) }
             .to raise_error error_class, error_message
         end
       end
 
       describe 'with an object' do
-        let(:object)      { Object.new.freeze }
-        let(:error_class) { Serializers::InvalidObjectError }
+        let(:object) { Object.new.freeze }
         let(:error_message) do
           "Unable to serialize #{object.inspect} with Spec::WidgetSerializer"
         end
 
-        it 'should raise an error' do
+        it 'should raise an invalid object error' do
           expect { serializer.call(object) }
-            .to raise_error error_class, error_message
+            .to raise_error Serializers::InvalidObjectError, error_message
         end
       end
 
       describe 'with a valid object' do
         let(:widget) { Spec::Widget.new }
-
-        it 'should not raise an error' do
-          expect { serializer.call(widget) }.not_to raise_error
-        end
-      end
-    end
-
-    context 'when there are many attributes' do
-      let(:described_class) { Spec::PersonSerializer }
-
-      example_class 'Spec::Person', Struct.new(:first_name, :last_name, :title)
-
-      # rubocop:disable RSpec/DescribedClass
-      example_class 'Spec::PersonSerializer', Serializers::BaseSerializer \
-      do |klass|
-        klass.attributes :first_name, :last_name, :title
-      end
-      # rubocop:enable RSpec/DescribedClass
-
-      describe 'with an object with empty properties' do
-        let(:object) { Spec::Person.new }
-
-        include_examples 'should serialize attributes',
-          :first_name,
-          :last_name,
-          :title
-      end
-
-      describe 'with an object with set properties' do
-        let(:object) { Spec::Person.new('Alan', 'Bradley', 'Programmer') }
-
-        include_examples 'should serialize attributes',
-          :first_name,
-          :last_name,
-          :title
-      end
-
-      context 'with a subclass' do
-        let(:described_class) { Spec::WorkerSerializer }
-
-        example_class 'Spec::Worker',
-          Struct.new(:first_name, :last_name, :title, :company)
-
-        example_class 'Spec::WorkerSerializer', 'Spec::PersonSerializer' \
-        do |klass|
-          klass.attribute :company
+        let(:error_message) do
+          Serializers::UndefinedSerializerError.message(widget)
         end
 
-        # rubocop:disable RSpec/NestedGroups
-        describe 'with an object with empty properties' do
-          let(:object) { Spec::Worker.new }
-
-          include_examples 'should serialize attributes',
-            :first_name,
-            :last_name,
-            :title,
-            :company
+        it 'should raise an undefined serializer error' do
+          expect { serializer.call widget }
+            .to raise_error Serializers::UndefinedSerializerError, error_message
         end
-
-        describe 'with an object with set properties' do
-          let(:object) do
-            Spec::Worker.new('Ed', 'Dillinger', 'Executive', 'Incom')
-          end
-
-          include_examples 'should serialize attributes',
-            :first_name,
-            :last_name,
-            :title,
-            :company
-        end
-        # rubocop:enable RSpec/NestedGroups
       end
     end
   end
